pyproject.toml
[tool.poetry]
name = "drf-dogs"
version = "0.1.0"
description = ""
authors = ["TheShadowfish <shadowfish@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.2"
psycopg2 = "^2.9.9"
python-decouple = "^3.8"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
